Fixed cartesius related problems   casting to long long in std::to_string call ...
Changed problem setup.

+++ ocean/Ocean.C
@@ -481,7 +481,7 @@ void Ocean::saveStateToFile(std::string const &filename)
 	EpetraExt::HDF5 HDF5(*comm_);
 	HDF5.Create(filename);
 	HDF5.Write("State", *state_);
-	HDF5.Write("Map-" +  std::to_string(comm_->NumProc()),
+	HDF5.Write("Map-" +  std::to_string((long long) comm_->NumProc()),
 			   *(domain_->GetSolveMap()));
 	HDF5.Write("Continuation parameter", "Value", parValue_);
 	
@@ -513,7 +513,7 @@ void Ocean::loadStateFromFile(std::string const &filename)
 	// Read map and state
 	HDF5.Open(filename);
 	
-	HDF5.Read("Map-" + std::to_string(comm_->NumProc()), map);
+	HDF5.Read("Map-" + std::to_string((long long) comm_->NumProc()), map);
 	HDF5.Read("State", *map, state);
 
 	// Obtain Epetra_Vector from multivector and construct state_
